feat: improved validator set keys ux
