+++ packages/page-staking-async/src/CommandCenter/index.tsx
@@ -10,8 +10,8 @@ import React, { useCallback, useEffect, useRef, useState } from 'react';
 
 import { Dropdown, styled } from '@polkadot/react-components';
 import { useApi } from '@polkadot/react-hooks';
+import { getApi } from '@polkadot/react-hooks/useStakingAsyncApis';
 
-import { getApi } from '../utils.js';
 import AssetHubSection from './ah.js';
 import RelaySection from './relay.js';
