+++ packages/page-staking-async/src/index.tsx
@@ -1,58 +1,18 @@
 // Copyright 2017-2025 @polkadot/app-staking-async authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { ApiPromise } from '@polkadot/api';
 import type { AppProps as Props } from '@polkadot/react-components/types';
 
-import React, { useEffect, useMemo, useState } from 'react';
+import React from 'react';
 
-import { createWsEndpoints } from '@polkadot/apps-config';
 import { styled } from '@polkadot/react-components';
-import { useApi } from '@polkadot/react-hooks';
+import { useStakingAsyncApis } from '@polkadot/react-hooks';
 
 import StakingRelayApp from './Relay/index.js';
 import StakingSystemApp from './System/index.js';
-import { getApi } from './utils.js';
-
-const allEndPoints = createWsEndpoints((k, v) => v?.toString() || k);
 
 function StakingApp ({ basePath, className = '', onStatusChange }: Props): React.ReactElement<Props> {
-  const { api, apiEndpoint } = useApi();
-
-  const [ahApi, setAhApi] = useState<ApiPromise>();
-  const [rcApi, setRcApi] = useState<ApiPromise>();
-
-  const isRelayChain = useMemo(() => !!api.tx.stakingAhClient, [api.tx.stakingAhClient]);
-
-  const rcEndPoints = useMemo(() => {
-    return (isRelayChain
-      ? apiEndpoint?.providers
-      : apiEndpoint?.valueRelay) || [];
-  }, [apiEndpoint?.providers, apiEndpoint?.valueRelay, isRelayChain]);
-
-  const ahEndPoints: string[] = useMemo(() => {
-    if (isRelayChain) {
-      return allEndPoints.find(({ genesisHashRelay, paraId }) =>
-        paraId === 1000 && genesisHashRelay === api.genesisHash.toHex()
-      )?.providers || [];
-    }
-
-    return apiEndpoint?.providers || [];
-  }, [api.genesisHash, apiEndpoint?.providers, isRelayChain]);
-
-  useEffect(() => {
-    if (isRelayChain) {
-      // Pick random endpoint
-      const ahUrl = ahEndPoints.at(Math.floor(Math.random() * ahEndPoints.length));
-
-      !!ahUrl && getApi(ahUrl).then((ahApi) => setAhApi(ahApi)).catch(console.log);
-    } else {
-      // Pick random endpoint
-      const rcUrl = rcEndPoints.at(Math.floor(Math.random() * rcEndPoints.length));
-
-      !!rcUrl && getApi(rcUrl).then((rcApi) => setRcApi(rcApi)).catch(console.log);
-    }
-  }, [ahEndPoints, isRelayChain, rcEndPoints]);
+  const { ahApi, ahEndPoints, isRelayChain, rcApi, rcEndPoints } = useStakingAsyncApis();
 
   return (
     <StyledMain className={`${className} staking--App`}>
